2QRB refactor as node
