+++ Quantum-Control-Applications-QuAM/Superconducting/.gitignore
@@ -1,1 +1,2 @@
 quam_libs/serialized_qua/
+calibration_graph/sync_hook.py
