--- Quantum-Control-Applications-QuAM/Superconducting/random_exp/2QRB/post_hook.py
@@ -1,1 +1,0 @@
-print("Post-hook executed!")
