--- Quantum-Control-Applications-QuAM/Superconducting/random_exp/2QRB/pre_hook.py
@@ -1,1 +1,0 @@
-print("Pre-hook executed!")
