--- Quantum-Control-Applications-QuAM/Superconducting/random_exp/2QRB/sync_hook.py
@@ -1,15 +1,0 @@
-from iqcc_cloud_client.runtime import get_qm_job
-from qm.qua import *
-
-job = get_qm_job()
-result = job.result_handles
-
-# The actual values from circuits_as_ints_batched
-circuits_as_ints_batched = [
-    [66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, 6, 56, 61, 65, 40, 6, 47, 66, 53, 0, 38, 65, 38, 0, 55, 66, 46, 8, 60, 65, 44, 8, 62, 66, 46, 0, 38, 65, 38, 0, 62, 66, 4, 48, 61, 65, 48, 4, 47, 66, 40, 6, 55, 65, 54, 0, 61, 66, 6, 56, 60, 65, 44, 0, 46, 66, 48, 7, 47, 65, 5, 48, 61, 66, 60, 56, 61, 65, 56, 60, 66, 41, 6, 55, 65, 49, 6, 66, 54, 56, 65, 37, 56, 62, 66, 5, 55, 65, 53, 7, 47, 66, 52, 56, 65, 37, 56, 60, 66, 38, 0, 44, 65, 4, 32, 62, 66, 46, 65, 44, 66, 44, 0, 47, 65, 5, 56, 60, 66, 49, 4, 47, 65, 1, 52, 66, 38, 56, 65, 53, 56, 62, 66, 52, 56, 62, 65, 38, 56, 60, 66, 32, 5, 39, 65, 32, 5, 39, 66, 61, 56, 61, 65, 54, 0, 61, 65, 46, 0, 54, 66, 4, 56, 60, 65, 1, 38, 65, 13, 32, 62, 66, 47, 15, 65, 56, 61, 65, 40, 13, 66, 36, 0, 61, 65, 33, 4, 47, 65, 52, 0, 62, 66, 44, 0, 62, 65, 44, 56, 62, 65, 5, 57, 66, 0, 52, 65, 4, 65, 48, 5, 55, 66, 48, 6, 65, 41, 6, 55, 65, 56, 66, 45, 1, 39, 65, 57, 65, 37, 7, 66, 45, 0, 62, 65, 12, 32, 62, 65, 49, 4, 47, 66, 36, 0, 53, 65, 45, 0, 62, 65, 37, 1, 66, 52, 65, 44, 0, 54, 65, 54, 0, 39, 66, 5, 48, 62, 65, 45, 8, 62, 65, 37, 0, 62, 66, 9, 39, 65, 46, 0, 39, 65, 48, 6, 55, 66, 37, 0, 37, 65, 44, 0, 55, 65, 6, 40, 60, 66, 37, 57, 65, 12, 48, 65, 40, 14, 66, 0, 36, 65, 13, 65, 38, 0, 61, 66, 54, 7, 47, 65, 12, 48, 62, 65, 54, 0, 37, 66, 44, 8, 61, 65, 46, 0, 62, 65, 5, 56, 66, 0, 36, 65, 12, 65, 32, 5, 66, 5, 41, 65, 14, 56, 61, 65, 45, 0, 60, 66, 44, 0, 52, 65, 53, 65, 48, 7, 65, 40, 60, 65, 53, 0, 54, 66, 36, 0, 61, 65, 41, 6, 65, 4, 32, 65, 55, 65, 41, 4, 55, 66, 53, 0, 46, 65, 45, 8, 62, 65, 52, 0, 46, 65, 61, 65, 48, 5, 66, 8, 52, 65, 37, 1, 39, 65, 37, 0, 45, 65, 5, 47, 65, 44, 0, 62, 66, 49, 6, 65, 44, 0, 61, 65, 8, 61, 65, 53, 1, 55, 65, 37, 7, 47, 66, 6, 32, 61, 65, 41, 65, 4, 56, 61, 65, 13, 56, 62, 65, 53, 0, 38, 66, 40, 12, 65, 41, 4, 55, 65, 44, 0, 39, 65, 12, 48, 61, 65, 33, 4, 55, 66, 49, 62, 65, 0, 36, 65, 46, 0, 36, 65, 40, 7, 55, 65, 37, 56, 66, 48, 5, 55, 65, 52, 0, 45, 65, 41, 6, 39, 65, 54, 0, 55, 65, 46, 8, 61, 66, 4, 32, 61, 65, 52, 0, 46, 65, 12, 32, 60, 65, 40, 60, 65, 53, 57, 66, 13, 65, 53, 0, 47, 65, 32, 4, 65, 36, 7, 47, 65, 61, 57, 66, 6, 32, 65, 6, 48, 60, 65, 38, 56, 60, 65, 46, 15, 65, 54, 0, 52, 66, 33, 4, 65, 45, 0, 37, 65, 32, 60, 65, 53, 65, 57, 66, 37, 0, 60, 65, 38, 0, 45, 65, 6, 48, 62, 65, 8, 61, 65, 46, 0, 53, 66, 53, 0, 47, 65, 53, 1, 55, 65, 56, 60, 65, 0, 61, 65, 13, 56, 60, 66, 14, 65, 6, 65, 45, 56, 61, 65, 45, 65, 5, 56, 61, 66, 4, 40, 60, 65, 54, 0, 45, 65, 52, 56, 60, 65, 14, 56, 61, 65, 47, 7, 55, 66, 6, 56, 61, 65, 52, 0, 61, 65, 53, 0, 39, 65, 54, 0, 55, 65, 52, 7, 39, 66, 4, 48, 61, 65, 14, 56, 61, 65, 57, 65, 48, 7, 65, 62, 56, 62, 66, 36, 7, 65, 1, 47, 65, 13, 57, 65, 54, 0, 38, 65, 38, 0, 66, 4, 40, 62, 65, 0, 61, 65, 32, 4, 39, 65, 36, 0, 36, 65, 0, 45, 65, 14, 32, 60, 65, 32, 5, 47, 65, 49, 7, 39, 65, 36, 0, 55, 66, 53, 0, 37, 65, 4, 32, 60, 65, 5, 57, 65, 14, 32, 60, 65, 46, 0, 52, 65, 54, 0, 54, 65, 13, 32, 60, 65, 54, 65, 41, 66, 52, 0, 39, 65, 44, 0, 44, 65, 46, 8, 61, 65, 5, 47, 65, 52, 56, 61, 65, 48, 4, 39, 65, 36, 7, 39, 65, 37, 56, 61, 65, 49, 6, 55, 66, 38, 0, 36, 65, 53, 0, 36, 65, 55, 7, 55, 65, 6, 32, 62, 65, 45, 0, 39, 65, 36, 7, 47, 65, 5, 40, 60, 65, 7, 39, 65, 54, 0, 45, 66, 14, 56, 62, 65, 48, 6, 39, 65, 38, 0, 47, 65, 4, 32, 62, 65, 37, 0, 60, 65, 48, 7, 65, 40, 5, 65, 13, 32, 65, 36, 56, 60, 66, 1, 47, 65, 45, 1, 65, 62, 56, 60, 65, 56, 65, 6, 32, 65, 33, 4, 55, 65, 54, 56, 62, 65, 0, 53, 65, 36, 0, 60, 66, 45, 0, 54, 65, 8, 60, 65, 41, 7, 55, 65, 44, 0, 53, 65, 38, 65, 45, 56, 61, 65, 37, 56, 65, 55, 7, 47, 65, 5, 32, 61, 66, 7, 65, 45, 7, 39, 65, 52, 0, 37, 65, 44, 0, 46, 65, 5, 39, 65, 33, 7, 65, 32, 4, 39, 65, 36, 56, 61, 65, 8, 37, 66, 1, 54, 65, 45, 1, 55, 65, 12, 32, 62, 65, 54, 7, 39, 65, 48, 61, 65, 46, 7, 55, 65, 0, 45, 65, 44, 0, 39, 65, 46, 0, 39, 66, 8, 54, 65, 32, 5, 47, 65, 15, 55, 65, 13, 65, 41, 4, 39, 65, 61, 56, 65, 40, 6, 47, 65, 45, 1, 65, 44, 0, 52, 66, 37, 7, 39, 65, 44, 0, 46, 65, 36, 56, 61, 65, 49, 6, 65, 4, 56, 62, 65, 53, 0, 47, 65, 37, 0, 46, 65, 37, 1, 65, 32, 6, 66, 53, 0, 52, 65, 45, 0, 37, 65, 32, 4, 47, 65, 53, 0, 61, 65, 37, 56, 65, 15, 65, 36, 0, 47, 65, 41, 62, 65, 40, 4, 55, 66, 6, 32, 65, 52, 7, 47, 65, 32, 5, 65, 41, 12, 65, 54, 0, 62, 65, 52, 0, 47, 65, 7, 39, 65, 45, 0, 65, 36, 0, 52, 66, 53, 0, 47, 65, 48, 4, 39, 65, 45, 0, 46, 65, 36, 0, 54, 65, 65, 6, 32, 60, 65, 40, 4, 39, 65, 44, 0, 46, 65, 40, 4, 55, 66, 36, 7, 65, 4, 56, 61, 65, 46, 0, 44, 65, 1, 65, 0, 36, 65, 44, 0, 45, 65, 46, 8, 61, 65, 38, 0, 54, 65, 33, 62, 66, 1, 39, 65, 38, 0, 53, 65, 8, 37, 65, 37, 0, 54, 65, 0, 44, 65, 40, 6, 39, 65, 38, 7, 47, 65, 32, 60, 65, 4, 56, 66, 38, 0, 60, 65, 32, 5, 65, 46, 7, 65, 44, 8, 61, 65, 6, 56, 60, 65, 46, 56, 61, 65, 6, 48, 61, 65, 9, 65, 44, 0, 39, 66, 33, 4, 39, 65, 9, 38, 65, 14, 32, 61, 65, 40, 5, 65, 60, 56, 65, 1, 60, 65, 8, 55, 65, 45, 0, 54, 65, 45, 8, 61, 66, 62, 56, 65, 33, 60, 65, 49, 4, 47, 65, 40, 4, 47, 65, 44, 56, 65, 41, 65, 9, 38, 65, 46, 7, 65, 45, 0, 38, 66, 46, 0, 44, 65, 54, 0, 53, 65, 5, 41, 65, 13, 48, 65, 9, 39, 65, 54, 7, 39, 65, 52, 0, 55, 65, 36, 65, 61, 66, 44, 0, 54, 65, 5, 32, 61, 65, 49, 6, 47, 65, 33, 7, 47, 65, 8, 62, 65, 38, 0, 38, 65, 48, 7, 47, 65, 40, 61, 65, 14, 48, 61, 65, 14, 55, 65, 47, 7, 47, 65, 48, 60, 65, 14, 48, 65, 6, 65, 33, 6, 55, 65, 41, 6, 39, 65, 41, 7, 55, 66, 7, 39, 65, 53, 1, 55, 65, 54, 0, 37, 65, 13, 56, 62, 65, 36, 0, 39, 65, 6, 65, 5, 33, 65, 52, 7, 39, 65, 48, 7, 39, 65, 52, 7, 39, 65, 45, 0, 46, 65, 14, 56, 62, 65, 36, 0, 54, 65, 48, 5, 65, 46, 0, 52, 65, 45, 7, 65, 1, 55, 66, 44, 0, 44, 65, 40, 4, 65, 12, 32, 62, 65, 0, 61, 65, 46, 65, 44, 0, 61, 65, 12, 48, 62, 65, 54, 56, 61, 65, 6, 56, 62, 65, 40, 6, 39, 65, 40, 14, 65, 8, 62, 65, 44, 56, 60, 65, 5, 56, 65, 46, 0, 44, 65, 37, 0, 46, 65, 6, 32, 62, 66, 36, 0, 55, 65, 13, 56, 65, 33, 4, 55, 65, 0, 39, 65, 4, 32, 62, 65, 55, 65, 37, 1, 47, 65, 49, 7, 65, 44, 0, 46, 65, 41, 12, 65, 45, 1, 47, 65, 60, 56, 62, 65, 36, 7, 55, 65, 37, 7, 65, 54, 7, 47, 65, 60, 56, 61, 65, 40, 61, 66, 48, 6, 55, 65, 14, 32, 65, 52, 56, 60, 65, 44, 0, 53, 65, 41, 7, 47, 65, 38, 0, 61, 65, 44, 0, 53, 65, 5, 39, 65, 33, 6, 47, 65, 40, 5, 39, 65, 52, 0, 38, 65, 62, 56, 60, 65, 54, 65, 40, 4, 39, 65, 0, 47, 65, 57, 65, 38, 0, 38, 66, 45, 7, 47, 65, 14, 55, 65, 54, 0, 44, 65, 52, 0, 65, 13, 49, 65, 8, 62, 65, 53, 65, 37, 65, 38, 0, 46, 65, 62, 65, 46, 15, 65, 46, 7, 55, 65, 45, 7, 55, 65, 0, 62, 65, 8, 61, 65, 40, 7, 55, 65, 38, 0, 66, 44, 0, 55, 65, 45, 56, 65, 46, 0, 47, 65, 5, 40, 62, 65, 62, 65, 4, 32, 65, 39, 7, 39, 65, 41, 65, 48, 7, 47, 65, 13, 56, 65, 48, 62, 65, 55, 65, 44, 0, 60, 65, 45, 1, 55, 65, 36, 0, 62, 65, 13, 32, 60, 65, 44, 0, 45, 66, 7, 65, 36, 0, 39, 65, 5, 48, 62, 65, 33, 7, 47, 65, 52, 0, 45, 65, 1, 60, 65, 13, 48, 60, 65, 38, 56, 65, 13, 65, 45, 0, 62, 65, 12, 56, 61, 65, 45, 7, 47, 65, 44, 0, 55, 65, 5, 32, 60, 65, 33, 62, 65, 4, 40, 60, 65, 0, 45, 66, 62, 56, 60, 65, 54, 56, 60, 65, 53, 0, 46, 65, 44, 0, 62, 65, 52, 0, 39, 65, 9, 38, 65, 33, 60, 65, 49, 6, 65, 37, 0, 61, 65, 4, 55, 65, 56, 61, 65, 46, 56, 62, 65, 13, 48, 62, 65, 14, 32, 61, 65, 32, 7, 55, 65, 14, 56, 65, 54, 7, 47, 66, 12, 65, 32, 65, 36, 56, 61, 65, 57, 65, 45, 0, 47, 65, 8, 39, 65, 49, 7, 39, 65, 13, 56, 65, 54, 56, 62, 65, 37, 1, 47, 65, 65, 46, 8, 62, 65, 55, 7, 55, 65, 53, 0, 61, 65, 38, 0, 46, 65, 13, 49, 65, 12, 66, 38, 0, 62, 65, 37, 0, 36, 65, 65, 12, 55, 65, 46, 56, 60, 65, 0, 60, 65, 57, 62, 65, 14, 55, 65, 8, 39, 65, 46, 8, 60, 65, 39, 7, 55, 65, 40, 6, 65, 33, 6, 65, 5, 48, 60, 65, 54, 7, 65, 4, 56, 65, 48, 7, 66, 52, 0, 38, 65, 54, 7, 55, 65, 45, 0, 44, 65, 53, 56, 60, 65, 33, 6, 47, 65, 5, 40, 60, 65, 6, 32, 65, 38, 0, 46, 65, 40, 14, 65, 33, 62, 65, 4, 56, 61, 65, 48, 6, 39, 65, 15, 55, 65, 53, 56, 60, 65, 46, 0, 54, 65, 49, 6, 55, 65, 38, 66, 9, 55, 65, 4, 39, 65, 4, 55, 65, 36, 0, 52, 65, 45, 0, 53, 65, 40, 5, 55, 65, 4, 40, 65, 4, 40, 65, 14, 56, 62, 65, 40, 14, 65, 33, 6, 65, 41, 7, 47, 65, 37, 56, 61, 65, 46, 8, 60, 65, 41, 14, 65, 14, 56, 65, 37, 7, 47, 66, 38, 0, 60, 65, 38, 0, 61, 65, 36, 0, 62, 65, 33, 6, 47, 65, 55, 7, 65, 46, 0, 44, 65, 49, 60, 65, 4, 40, 61, 65, 0, 55, 65, 38, 0, 39, 65, 62, 56, 62, 65, 48, 4, 47, 65, 5, 33, 65, 6, 56, 60, 65, 52, 0, 52, 65, 45, 65, 44, 0, 38, 66, 4, 56, 62, 65, 1, 62, 65, 14, 48, 65, 6, 48, 61, 65, 7, 39, 65, 40, 62, 65, 41, 6, 47, 65, 32, 6, 65, 61, 56, 62, 65, 8, 55, 65, 49, 4, 39, 65, 54, 7, 65, 48, 7, 47, 65, 5, 40, 61, 65, 52, 0, 39, 65, 32, 5, 39, 65, 32, 5, 39, 66, 9, 36, 65, 53, 56, 61, 65, 49, 65, 60, 65, 12, 56, 61, 65, 54, 0, 45, 65, 44, 0, 45, 65, 57, 65, 5, 57, 65, 4, 40, 65, 32, 7, 47, 65, 49, 7, 47, 65, 41, 4, 47, 65, 37, 65, 36, 7, 39, 65, 9, 38, 65, 36, 0, 62, 66, 4, 48, 65, 13, 56, 65, 61, 56, 60, 65, 40, 5, 55, 65, 53, 0, 65, 32, 4, 47, 65, 36, 0, 36, 65, 9, 55, 65, 36, 0, 54, 65, 40, 65, 44, 0, 45, 65, 1, 54, 65, 37, 0, 62, 65, 52, 0, 47, 65, 32, 7, 39, 65, 54, 0, 45, 65, 53, 1, 55, 66, 45, 7, 39, 65, 12, 32, 62, 65, 14, 56, 60, 65, 1, 39, 65, 62, 56, 62, 65, 37, 56, 60, 65, 52, 7, 55, 65, 52, 7, 65, 4, 56, 61, 65, 0, 62, 65, 13, 33, 65, 53, 56, 62, 65, 52, 0, 54, 65, 52, 0, 44, 65, 46, 0, 60, 65, 44, 0, 60, 65, 9, 62, 66, 45, 1, 47, 65, 54, 0, 61, 65, 52, 0, 65, 55, 7, 39, 65, 0, 54, 65, 52, 0, 44, 65, 44, 0, 52, 65, 54, 56, 62, 65, 52, 7, 55, 65, 46, 8, 60, 65, 45, 0, 60, 65, 1, 47, 65, 52, 0, 54, 65, 44, 0, 47, 65, 41, 15, 65, 61, 56, 61, 65, 33, 4, 47, 66, 40, 5, 47, 65, 45, 0, 53, 65, 54, 56, 61, 65, 44, 56, 62, 65, 44, 8, 62, 65, 0, 37, 65, 8, 61, 65, 36, 0, 47, 65, 5, 48, 61, 65, 62, 56, 62, 65, 33, 4, 47, 65, 32, 62, 65, 44, 7, 55, 65, 52, 0, 61, 65, 61, 56, 62, 65, 13, 32, 62, 65, 66, 8, 61, 65, 36, 0, 53, 65, 37, 0, 53, 65, 41, 7, 65, 53, 1, 65, 53, 1, 39, 65, 40, 15, 65, 44, 0, 44, 65, 12, 48, 60, 65, 33, 7, 47, 65, 52, 65, 53, 0, 37, 65, 5, 48, 61, 65, 38, 0, 39, 65, 0, 38, 65, 6, 32, 65, 48, 5, 65, 37, 0, 54, 65, 40, 65, 45, 8, 60, 65, 37, 0, 39, 65, 52, 0, 47, 65, 45, 7, 65, 44, 56, 62, 65, 0, 47, 65, 6, 48, 61, 65, 44, 0, 53, 65, 4, 40, 65, 13, 55, 65, 33, 4, 65, 53, 1, 65, 38, 0, 54, 65, 53, 0, 45, 66, 12, 56, 60, 65, 38, 0, 52, 65, 32, 65, 46, 65, 33, 65, 54, 56, 60, 65, 14, 32, 61, 65, 36, 0, 38, 65, 45, 1, 65, 12, 56, 65, 45, 0, 47, 65, 7, 55, 65, 1, 39, 65, 54, 7, 65, 14, 32, 62, 65, 33, 6, 47, 65, 1, 52, 65, 41, 60, 65, 52, 65, 36, 56, 62, 65, 9, 38, 65, 38, 56, 61, 65, 7, 39, 65, 55, 7, 65, 48, 7, 55, 65, 53, 7, 65, 14, 32, 61, 65, 8, 38, 65, 36, 0, 38, 65, 44, 0, 52, 65, 54, 0, 38, 65, 33, 6, 47, 65, 46, 0, 55, 66, 12, 32, 60, 65, 9, 65, 48, 4, 55, 65, 12, 48, 60, 65, 5, 32, 60, 65, 36, 0, 65, 36, 65, 12, 48, 65, 44, 7, 47, 65, 12, 32, 61, 65, 57, 60, 65, 52, 0, 46, 65, 49, 7, 39, 65, 6, 65, 41, 12, 65, 33, 7, 55, 65, 37, 0, 47, 65, 48, 5, 47, 65, 62, 56, 60, 65, 9, 39, 65, 41, 4, 39, 65, 6, 56, 65, 36, 56, 60, 65, 40, 7, 39, 65, 44, 8, 60, 65, 6, 40, 61, 65, 49, 4, 47, 65, 4, 32, 62, 65, 45, 7, 65, 38, 0, 53, 65, 4, 32, 62, 65, 33, 4, 65, 60, 66, 1, 60, 65, 53, 56, 61, 65, 12, 32, 61, 65, 32, 6, 47, 65, 52, 0, 55, 65, 49, 4, 55, 65, 37, 7, 47, 65, 45, 0, 60, 65, 56, 60, 65, 53, 0, 61, 65, 47, 15, 65, 41, 7, 47, 65, 37, 56, 65, 49, 7, 55, 65, 53, 56, 65, 40, 60, 65, 6, 56, 65, 45, 8, 65, 53, 1, 65, 0, 46, 65, 33, 6, 55, 65, 41, 4, 47, 65, 52, 0, 52, 65, 0, 36, 65, 8, 55, 65, 46, 56, 62, 65, 49, 4, 47, 65, 4, 32, 62, 65, 7, 39, 65, 33, 4, 47, 65, 6, 40, 65, 33, 6, 55, 65, 46, 8, 61, 66, 13, 48, 62, 65, 44, 8, 61, 65, 45, 56, 62, 65, 4, 32, 60, 65, 1, 52, 65, 47, 7, 55, 65, 4, 47, 65, 5, 55, 65, 15, 39, 65, 4, 40, 60, 65, 49, 7, 55, 65, 5, 55, 65, 6, 40, 65, 52, 0, 60, 65, 13, 49, 65, 45, 0, 54, 65, 44, 0, 53, 65, 48, 7, 47, 65, 14, 39, 65, 49, 7, 39, 65, 40, 14, 65, 49, 60, 65, 44, 8, 60, 65, 6, 32, 65, 38, 0, 65, 62, 56, 60, 65, 40, 60, 65, 52, 7, 39, 65, 54, 0, 39, 65, 54, 0, 36, 65, 13, 55, 65, 6, 56, 61, 65, 36, 0, 37, 66, 4, 56, 65, 48, 65, 33, 6, 47, 65, 53, 0, 44, 65, 54, 0, 45, 65, 12, 48, 60, 65, 0, 60, 65, 32, 6, 65, 38, 7, 47, 65, 40, 14, 65, 56, 60, 65, 5, 39, 65, 37, 56, 62, 65, 4, 48, 60, 65, 46, 56, 60, 65, 32, 5, 65, 6, 48, 61, 65, 37, 0, 61, 65, 12, 32, 65, 44, 8, 60, 65, 0, 52, 65, 0, 60, 65, 37, 0, 52, 65, 36, 7, 47, 65, 4, 40, 62, 65, 36, 0, 44, 65, 47, 7, 65, 45, 56, 62, 65, 13, 48, 61, 65, 45, 0, 62, 65, 0, 44, 65, 46, 8, 62, 65, 55, 7, 55, 66, 40, 6, 65, 56, 61, 65, 45, 56, 62, 65, 5, 48, 65, 46, 0, 37, 65, 62, 56, 61, 65, 46, 7, 65, 4, 40, 62, 65, 55, 65, 40, 4, 47, 65, 36, 0, 61, 65, 12, 56, 60, 65, 14, 65, 54, 0, 46, 65, 44, 0, 65, 46, 0, 55, 65, 14, 48, 65, 46, 0, 61, 65, 45, 57, 65, 6, 56, 62, 65, 53, 0, 54, 65, 47, 7, 65, 36, 0, 38, 65, 9, 52, 65, 49, 4, 65, 0, 53, 65, 44, 0, 54, 65, 37, 0, 36, 65, 4, 55, 65, 46, 0, 65, 40, 5, 55, 65, 53, 1, 65, 6, 40, 60, 66, 49, 6, 47, 65, 8, 38, 65, 48, 4, 47, 65, 4, 40, 60, 65, 40, 62, 65, 45, 0, 62, 65, 13, 56, 62, 65, 44, 7, 55, 65, 53, 0, 61, 65, 13, 32, 65, 38, 7, 65, 6, 32, 60, 65, 41, 7, 55, 65, 12, 32, 60, 65, 36, 0, 44, 65, 37, 0, 65, 6, 40, 60, 65, 5, 40, 65, 0, 53, 65, 55, 7, 65, 45, 0, 38, 65, 4, 56, 62, 65, 44, 7, 55, 65, 44, 0, 62, 65, 1, 39, 65, 45, 0, 53, 65, 54, 7, 55, 65, 12, 65, 54, 0, 47, 65, 8, 37, 65, 36, 0, 61, 65, 49, 65, 41, 7, 55, 66, 15, 39, 65, 45, 56, 65, 38, 0, 38, 65, 45, 0, 39, 65, 46, 56, 62, 65, 53, 0, 46, 65, 13, 56, 60, 65, 47, 65, 1, 46, 65, 38, 0, 54, 65, 13, 57, 65, 52, 56, 61, 65, 33, 4, 47, 65, 46, 8, 65, 53, 0, 46, 65, 37, 56, 62, 65, 14, 65, 57, 60, 65, 1, 47, 65, 4, 56, 61, 65, 45, 56, 62, 65, 37, 0, 52, 65, 49, 7, 39, 65, 52, 0, 38, 65, 44, 15, 65, 37, 1, 65, 4, 39, 65, 0, 39, 65, 36, 56, 61, 65, 13, 56, 60, 65, 52, 0, 36, 65, 45, 7, 65, 36, 0, 45, 66, 8, 39, 65, 4, 40, 61, 65, 38, 0, 45, 65, 5, 40, 65, 4, 32, 61, 65, 39, 7, 39, 65, 13, 32, 65, 38, 7, 55, 65, 45, 56, 65, 48, 7, 39, 65, 45, 0, 61, 65, 38, 56, 60, 65, 44, 0, 38, 65, 53, 7, 39, 65, 54, 7, 47, 65, 9, 60, 65, 13, 55, 65, 44, 7, 65, 46, 0, 36, 65, 12, 56, 61, 65, 36, 7, 39, 65, 13, 32, 61, 65, 37, 0, 52, 65, 6, 47, 65, 38, 0, 36, 65, 47, 7, 47, 65, 55, 7, 39, 65, 46, 15, 65, 61, 56, 61, 65, 45, 0, 47, 65, 6, 48, 61, 65, 45, 0, 60, 65, 5, 56, 60, 66, 6, 48, 60, 65, 53, 0, 55, 65, 45, 9, 65, 38, 56, 62, 65, 36, 0, 60, 65, 40, 60, 65, 56, 62, 65, 36, 0, 47, 65, 8, 61, 65, 38, 56, 65, 45, 0, 47, 65, 9, 36, 65, 37, 56, 61, 65, 32, 5, 55, 65, 41, 7, 65, 37, 65, 46, 0, 45, 65, 37, 0, 53, 65, 47, 15, 65, 45, 8, 61, 65, 49, 7, 39, 65, 38, 0, 62, 65, 52, 0, 55, 65, 45, 0, 62, 65, 54, 0, 65, 9, 52, 65, 33, 6, 65, 37, 7, 47, 65, 6, 55, 65, 48, 4, 55, 65, 44, 56, 60, 65, 44, 0, 54, 65, 41, 6, 55, 66, 37, 0, 60, 65, 41, 65, 13, 56, 62, 65, 37, 0, 55, 65, 0, 47, 65, 6, 32, 62, 65, 36, 0, 38, 65, 46, 56, 60, 65, 44, 0, 46, 65, 13, 32, 62, 65, 37, 56, 62, 65, 52, 0, 47, 65, 14, 55, 65, 52, 7, 39, 65, 44, 7, 65, 54, 0, 47, 65, 40, 4, 47, 65, 53, 0, 46, 65, 45, 7, 55, 65, 39, 7, 39, 65, 13, 57, 65, 6, 40, 60, 65, 40, 5, 65, 5, 57, 65, 55, 65, 48, 4, 65, 6, 48, 65, 4, 40, 60, 65, 52, 7, 65, 53, 0, 60, 65, 37, 56, 65, 44, 0, 46, 65, 60, 56, 60, 66, 4, 32, 62, 65, 52, 65, 44, 0, 52, 65, 53, 65, 45, 0, 53, 65, 52, 0, 60, 65, 8, 65, 53, 0, 36, 65, 40, 5, 47, 65, 44, 0, 45, 65, 45, 1, 55, 65, 54, 0, 47, 65, 40, 60, 65, 5, 32, 60, 65, 38, 0, 38, 65, 32, 65, 53, 0, 39, 65, 0, 53, 65, 46, 8, 65, 54, 56, 61, 65, 41, 4, 55, 65, 4, 48, 61, 65, 33, 62, 65, 38, 0, 60, 65, 45, 0, 52, 65, 4, 32, 65, 40, 61, 65, 40, 7, 39, 65, 53, 0, 61, 65, 38, 7, 65, 48, 6, 47, 65, 40, 6, 47, 65, 52, 0, 39, 66, 53, 7, 39, 65, 37, 0, 37, 65, 37, 57, 65, 44, 0, 37, 65, 9, 38, 65, 46, 56, 62, 65, 36, 0, 39, 65, 54, 0, 45, 65, 49, 6, 39, 65, 8, 38, 65, 40, 5, 55, 65, 52, 0, 46, 65, 46, 0, 52, 65, 48, 4, 55, 65, 48, 6, 65, 48, 62, 65, 6, 55, 65, 12, 48, 60, 65, 4, 56, 60, 65, 37, 7, 39, 65, 4, 65, 41, 7, 65, 6, 32, 62, 65, 54, 0, 54, 65, 37, 1, 55, 65, 13, 32, 60, 65, 13, 32, 60, 65, 37, 0, 53, 65, 54, 0, 46, 65, 44, 0, 62, 65, 32, 4, 39, 65, 54, 0, 52, 65, 46, 56, 62, 66, 61, 65, 46, 7, 39, 65, 44, 0, 37, 65, 52, 0, 44, 65, 49, 65, 33, 4, 47, 65, 33, 4, 47, 65, 49, 62, 65, 41, 6, 55, 65, 37, 0, 55, 65, 4, 56, 60, 65, 45, 0, 47, 65, 53, 0, 44, 65, 54, 56, 62, 65, 54, 56, 60, 65, 6, 56, 61, 65, 6, 48, 65, 46, 0, 55, 65, 14, 55, 65, 54, 0, 47, 65, 12, 56, 65, 38, 0, 54, 65, 38, 0, 62, 65, 38, 0, 36, 65, 4, 32, 60, 65, 4, 56, 60, 65, 41, 4, 47, 65, 45, 7, 47, 65, 60, 56, 65, 1, 55, 65, 38, 0, 54, 65, 12, 56, 61, 65, 12, 55, 66, 46, 0, 45, 65, 53, 0, 65, 6, 40, 65, 6, 48, 65, 60, 56, 65, 5, 39, 65, 44, 7, 65, 33, 6, 65, 37, 0, 53, 65, 5, 32, 62, 65, 44, 0, 65, 5, 32, 61, 65, 36, 0, 39, 65, 49, 62, 65, 37, 0, 65, 54, 0, 62, 65, 46, 0, 39, 65, 12, 32, 65, 49, 60, 65, 32, 6, 65, 41, 7, 39, 65, 52, 7, 47, 65, 13, 55, 65, 33, 6, 47, 65, 0, 37, 65, 7, 39, 65, 41, 60, 65, 37, 0, 54, 65, 52, 0, 39, 65, 6, 47, 65, 54, 0, 55, 65, 44, 7, 55, 65, 0, 55, 66, 4, 39, 65, 48, 4, 47, 65, 62, 56, 60, 65, 36, 0, 54, 65, 53, 7, 47, 65, 14, 48, 61, 65, 44, 0, 47, 65, 48, 4, 47, 65, 37, 65, 5, 55, 65, 0, 55, 65, 32, 7, 65, 41, 6, 39, 65, 36, 7, 39, 65, 4, 56, 62, 65, 8, 62, 65, 54, 56, 65, 41, 15, 65, 1, 60, 65, 38, 0, 47, 65, 52, 0, 38, 65, 44, 56, 61, 65, 41, 7, 39, 65, 37, 0, 37, 65, 12, 65, 48, 7, 55, 65, 46, 0, 55, 65, 53, 56, 62, 65, 0, 46, 65, 49, 62, 65, 1, 55, 65, 54, 0, 65, 37, 0, 60, 66, 44, 0, 46, 65, 49, 6, 39, 65, 48, 60, 65, 60, 56, 62, 65, 53, 57, 65, 4, 47, 65, 53, 1, 65, 33, 4, 65, 54, 0, 47, 65, 13, 33, 65, 53, 1, 47, 65, 44, 7, 39, 65, 52, 65, 45, 1, 55, 65, 45, 0, 61, 65, 40, 4, 47, 65, 5, 49, 65, 32, 61, 65, 62, 56, 65, 53, 0, 60, 65, 45, 57, 65, 0, 45, 65, 0, 60, 65, 53, 56, 60, 65, 56, 62, 65, 6, 55, 65, 45, 9, 65, 48, 65, 32, 65, 62, 65, 45, 56, 61, 65, 37, 56, 62, 65, 6, 55, 66, 49, 62, 65, 37, 65, 9, 60, 65, 8, 38, 65, 38, 0, 65, 37, 0, 52, 65, 40, 4, 39, 65, 5, 48, 60, 65, 14, 48, 61, 65, 32, 4, 39, 65, 38, 0, 36, 65, 62, 56, 62, 65, 32, 4, 39, 65, 1, 36, 65, 62, 56, 61, 65, 0, 52, 65, 53, 0, 38, 65, 53, 0, 38, 65, 37, 56, 61, 65, 47, 7, 39, 65, 52, 0, 60, 65, 40, 7, 39, 65, 13, 48, 62, 65, 52, 0, 61, 65, 32, 5, 55, 65, 41, 7, 55, 65, 0, 53, 65, 52, 0, 37, 65, 46, 0, 38, 65, 12, 65, 5, 56, 62, 65, 46, 0, 37, 65, 5, 57, 66, 52, 0, 62, 65, 1, 47, 65, 9, 55, 65, 44, 8, 65, 8, 65, 5, 32, 62, 65, 53, 0, 53, 65, 33, 65, 38, 0, 52, 65, 38, 0, 36, 65, 57, 60, 65, 33, 4, 55, 65, 48, 7, 39, 65, 6, 40, 62, 65, 37, 0, 53, 65, 9, 55, 65, 4, 40, 60, 65, 1, 47, 65, 0, 45, 65, 33, 62, 65, 44, 8, 65, 14, 39, 65, 48, 7, 55, 65, 8, 36, 65, 6, 40, 60, 65, 46, 0, 53, 65, 36, 0, 52, 65, 44, 0, 37, 65, 0, 60, 65, 13, 56, 60, 65, 0, 44, 65, 0, 52, 65, 44, 56, 60, 66],
-]
-
-for id, batch in enumerate(circuits_as_ints_batched):
-    job.push_to_input_stream("sequence", batch)
-    result.measurements.wait_for_values(id+1)
-    print(f"{id}: Received ", str(result.measurements.fetch(id)))
